func on each arr element

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -12,10 +12,10 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-int x;
+unsigned int x;
 if (array != NULL || action != NULL)
 {
-for (x = 0; x < (int) size; x++)
+for (x = 0; x < size; x++)
 action(array[x]);
 }
 }
